+++ gubanov.ivan/T3/local.h
@@ -1,0 +1,16 @@
+#ifndef LOCAL_H
+#define LOCAL_H
+
+#include <iostream>
+#include <iterator>
+#include <vector>
+#include <algorithm>
+#include <functional>
+#include <fstream>
+#include <limits>
+#include <cmath>
+#include <numeric>
+#include <string>
+#include <iomanip>
+
+#endif
